legal-rag-chatbot/requirements.txt
streamlit==1.35.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
PyMuPDF==1.24.5
openai==1.35.0
python-dotenv==1.0.1
numpy==1.26.4
groq==0.9.0
